+++ src/main/java/racingcar/domain/dto/RequestAttemptDto.java
@@ -1,0 +1,16 @@
+package racingcar.domain.dto;
+
+import racingcar.domain.vo.Attempt;
+
+public class RequestAttemptDto {
+
+    private Attempt attempt;
+
+    public RequestAttemptDto(Attempt attempt) {
+        this.attempt = attempt;
+    }
+
+    public Attempt getAttempt() {
+        return attempt;
+    }
+}
