+++ src/main/java/racingcar/domain/dto/RequestCarsDto.java
@@ -1,0 +1,15 @@
+package racingcar.domain.dto;
+
+import racingcar.domain.vo.Cars;
+
+public class RequestCarsDto {
+    private Cars cars;
+
+    public RequestCarsDto(Cars cars) {
+        this.cars = cars;
+    }
+
+    public Cars getCars() {
+        return cars;
+    }
+}
