+++ src/main/java/racingcar/domain/dto/ResponseWinnersDto.java
@@ -1,0 +1,17 @@
+package racingcar.domain.dto;
+
+import java.util.List;
+import racingcar.domain.vo.Car;
+
+public class ResponseWinnersDto {
+
+    private List<Car> winners;
+
+    public ResponseWinnersDto(List<Car> winners) {
+        this.winners = winners;
+    }
+
+    public List<Car> getWinners() {
+        return winners;
+    }
+}
