+++ src/main/java/racingcar/domain/util/MovableNumberGenerator.java
@@ -1,0 +1,9 @@
+package racingcar.domain.util;
+
+public class MovableNumberGenerator implements NumberGenerator {
+
+    @Override
+    public int generate() {
+        return MovementUtil.getMoveCondition();
+    }
+}
