+++ src/main/java/racingcar/domain/util/MovementUtil.java
@@ -1,0 +1,17 @@
+package racingcar.domain.util;
+
+public class MovementUtil {
+
+    private static final int MOVE_CONDITION = 4;
+
+    public static boolean isMoveForward(int randomNumber) {
+        if (randomNumber >= MOVE_CONDITION) {
+            return true;
+        }
+        return false;
+    }
+
+    public static int getMoveCondition() {
+        return MOVE_CONDITION;
+    }
+}
