+++ src/main/java/racingcar/domain/util/NonMovableNumberGenerator.java
@@ -1,0 +1,9 @@
+package racingcar.domain.util;
+
+public class NonMovableNumberGenerator implements NumberGenerator {
+
+    @Override
+    public int generate() {
+        return MovementUtil.getMoveCondition() - 1;
+    }
+}
