+++ src/main/java/racingcar/domain/util/NumberGenerator.java
@@ -1,0 +1,6 @@
+package racingcar.domain.util;
+
+public interface NumberGenerator {
+
+    int generate();
+}
