--- src/main/java/racingcar/dto/RequestAttemptDto.java
@@ -1,16 +1,0 @@
-package racingcar.dto;
-
-import racingcar.vo.Attempt;
-
-public class RequestAttemptDto {
-
-    private Attempt attempt;
-
-    public RequestAttemptDto(Attempt attempt) {
-        this.attempt = attempt;
-    }
-
-    public Attempt getAttempt() {
-        return attempt;
-    }
-}
