--- src/main/java/racingcar/dto/ResponseWinnersDto.java
@@ -1,17 +1,0 @@
-package racingcar.dto;
-
-import java.util.List;
-import racingcar.vo.Car;
-
-public class ResponseWinnersDto {
-
-    private List<Car> winners;
-
-    public ResponseWinnersDto(List<Car> winners) {
-        this.winners = winners;
-    }
-
-    public List<Car> getWinners() {
-        return winners;
-    }
-}
