--- src/main/java/racingcar/util/MovableNumberGenerator.java
@@ -1,9 +1,0 @@
-package racingcar.util;
-
-public class MovableNumberGenerator implements NumberGenerator {
-
-    @Override
-    public int generate() {
-        return MovementUtil.getMoveCondition();
-    }
-}
