--- src/main/java/racingcar/util/NonMovableNumberGenerator.java
@@ -1,9 +1,0 @@
-package racingcar.util;
-
-public class NonMovableNumberGenerator implements NumberGenerator {
-
-    @Override
-    public int generate() {
-        return MovementUtil.getMoveCondition() - 1;
-    }
-}
