--- src/main/java/racingcar/util/NumberGenerator.java
@@ -1,6 +1,0 @@
-package racingcar.util;
-
-public interface NumberGenerator {
-
-    int generate();
-}
